refactor(path): Path 도메인 경로 수정 - 구현체와 인터페이스의 뒤바뀐 위치 수정
