+++ backend/src/main/java/groom/backend/domain/path/controller/PathController.java
@@ -2,11 +2,10 @@ package groom.backend.domain.path.controller;
 
 import groom.backend.domain.path.dto.request.PathFindRequest;
 import groom.backend.domain.path.dto.response.PathFindResponse;
-import groom.backend.domain.path.service.impl.PathService;
+import groom.backend.domain.path.service.spec.PathService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
-import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
